Migrate to Android Preference library
Preference library allows to create preferences UI in a simple way.
It is a standard Android component.

However, it doesn't follow all the best practices.
For example, it executes all IO logic on the main thread, and
since we've enabled strict mode, it crashes our app.
So, we have to handle all persistence logic ourselves.

Also, it's pain in the ass to customize this library.